refactor: Gender を riddle.py に統合
Gender と GenderType はなぞなぞ機能でのみ使用されるため、
valueobject/completion/gender.py を削除し、
valueobject/completion/riddle.py に統合。

- gender.py を削除
- riddle.py に Gender, GenderType を追加
- 関連するすべてのインポート文を更新

既存のテストがすべてパス (12 tests OK)
