--- llm_chat/domain/valueobject/completion/gender.py
@@ -1,16 +1,0 @@
-from dataclasses import dataclass
-from enum import Enum
-
-
-class GenderType(Enum):
-    MAN = "man"
-    WOMAN = "woman"
-
-
-@dataclass
-class Gender:
-    gender: GenderType
-
-    @property
-    def name(self) -> str:
-        return "男性" if self.gender == GenderType.MAN else "女性"
